made all Integer designations BigInteger

+++ TWITOFF/models.py
@@ -19,9 +19,11 @@ class User(DB.Model):
 
 class Tweet(DB.Model):
     """Twitter users' tweets for analysis"""
-    id = DB.Column(DB.Integer, primary_key=True)
+    # id = DB.Column(DB.Integer, primary_key=True)
+    id = DB.Column(DB.BigInteger, primary_key=True)
     text = DB.Column(DB.Unicode(300))
-    user_id = DB.Column(DB.Integer, DB.ForeignKey('user.id'), nullable=False)
+    # user_id = DB.Column(DB.Integer, DB.ForeignKey('user.id'), nullable=False)
+    user_id = DB.Column(DB.BigInteger, DB.ForeignKey('user.id'), nullable=False)
     user = DB.relationship('User', backref=DB.backref('tweets'), lazy=True)
 
     # Added for pickling data
